quartermaster-graph/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quartermaster-graph"
version = "0.1.0"
description = "Framework-agnostic agent graph schema for AI agent workflows as directed acyclic graphs"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "MindMade", email = "info@mindmade.io" },
]
keywords = ["ai", "agents", "graph", "dag", "workflow", "schema"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/quartermaster-ai/quartermaster"
Repository = "https://github.com/quartermaster-ai/quartermaster"
Issues = "https://github.com/quartermaster-ai/quartermaster/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/quartermaster_graph"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/quartermaster_graph --cov-report=term-missing"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
